优化代码

+++ xloglib/src/main/java/com/xuexiang/xlog/crash/CrashHandler.java
@@ -30,10 +30,11 @@ import java.io.File;
 import java.lang.Thread.UncaughtExceptionHandler;
 
 /**
- * 在Application中统一捕获异常，保存到文件中下次再打开时上传
- *
- * @author xuexiang
- * @date 2018/1/28 下午10:03
+ * <pre>
+ *     desc   : 在Application中统一捕获异常，保存到文件中下次再打开时上传
+ *     author : xuexiang
+ *     time   : 2018/5/13 下午10:33
+ * </pre>
  */
 public class CrashHandler implements UncaughtExceptionHandler, ICrashHandler {
     private final static String DEFAULT_CRASH_LOG_FLAG = "crash_log";
@@ -53,6 +54,8 @@ public class CrashHandler implements UncaughtExceptionHandler, ICrashHandler {
      * 崩溃监听
      */
     private OnCrashListener mOnCrashListener;
+
+    private OnAppExitListener mOnAppExitListener;
 
     /**
      * 崩溃是否处理完毕
@@ -142,6 +145,17 @@ public class CrashHandler implements UncaughtExceptionHandler, ICrashHandler {
         return this;
     }
 
+    /**
+     * 设置应用程序退出监听
+     *
+     * @param onAppExitListener
+     * @return
+     */
+    public CrashHandler setOnAppExitListener(OnAppExitListener onAppExitListener) {
+        mOnAppExitListener = onAppExitListener;
+        return this;
+    }
+
     @Override
     public CrashHandler setIsHandledCrash(boolean isHandled) {
         mIsHandledCrash = isHandled;
@@ -197,24 +211,29 @@ public class CrashHandler implements UncaughtExceptionHandler, ICrashHandler {
                 }
             }
 
-            if (mIsNeedReopen) {
-                Intent intent = new Intent(mContext, StartAppReceiver.class);
-                PendingIntent restartIntent = PendingIntent.getBroadcast(mContext, 0, intent, PendingIntent.FLAG_ONE_SHOT);
-                //退出程序
-                AlarmManager mgr = (AlarmManager) mContext.getSystemService(Context.ALARM_SERVICE);
-                mgr.set(AlarmManager.RTC, System.currentTimeMillis() + 1000, restartIntent); // 1秒钟后重启应用
+            if (mOnAppExitListener == null) {
+                if (mIsNeedReopen) {
+                    Intent intent = new Intent(mContext, StartAppReceiver.class);
+                    PendingIntent restartIntent = PendingIntent.getBroadcast(mContext, 0, intent, PendingIntent.FLAG_ONE_SHOT);
+                    //退出程序
+                    AlarmManager mgr = (AlarmManager) mContext.getSystemService(Context.ALARM_SERVICE);
+                    mgr.set(AlarmManager.RTC, System.currentTimeMillis() + 1000, restartIntent); // 1秒钟后重启应用
+                }
+
+                android.os.Process.killProcess(android.os.Process.myPid());
+                System.exit(0);
+            } else {
+                mOnAppExitListener.onAppExit();
             }
 
-            android.os.Process.killProcess(android.os.Process.myPid());
-            System.exit(0);
         }
     }
 
     /**
      * 自定义错误处理,收集错误信息 发送错误报告等操作均在此完成.
      *
-     * @param throwable
-     * @return true:如果处理了该异常信息;否则返回false.
+     * @param throwable 出错信息
+     * @return {@code true}:处理了该异常信息;<br>{@code false}:该异常信息未处理。
      */
     private boolean handleException(Throwable throwable) {
         if (throwable == null || mContext == null || mOnCrashListener == null) {
@@ -263,4 +282,13 @@ public class CrashHandler implements UncaughtExceptionHandler, ICrashHandler {
         return Log.getStackTraceString(throwable);
     }
 
+    /**
+     * 应用即将退出的监听
+     */
+    public interface OnAppExitListener {
+        /**
+         * 应用即将退出
+         */
+        void onAppExit();
+    }
 }
